+++ frontend/public/env-config.js
@@ -1,7 +1,7 @@
 window._env_ = {
-  REACT_APP_APPLICATION_DOMAIN: "dev-7ranai11.auth0.com",
-  REACT_APP_APPLICATION_CLIENT_ID: "318GitavTaWR7d17h4DKuoCme9VgjYDG",
-  REACT_APP_AUDIENCE: "https://opensource.legal/contracts",
+  REACT_APP_APPLICATION_DOMAIN: "opencontracts-prod.us.auth0.com",
+  REACT_APP_APPLICATION_CLIENT_ID: "A4PZxnRxvwbJ7Y7FgR5vbXo7euzyKKcz",
+  REACT_APP_AUDIENCE: "https://contracts.opensource.legal",
   REACT_APP_API_ROOT_URL: "http://localhost:8000",
   REACT_APP_USE_AUTH0: "true",
   REACT_APP_USE_ANALYZERS: "",